+++ lib/grape/parser/json.rb
@@ -4,6 +4,9 @@ module Grape
       class << self
         def call(object, env)
           MultiJson.load(object)
+        rescue MultiJson::ParseError
+          # handle JSON parsing errors via the rescue handlers or provide error message
+          raise Grape::Exceptions::InvalidMessageBody, 'application/json'
         end
       end
     end