+++ lib/grape/parser/xml.rb
@@ -4,6 +4,9 @@ module Grape
       class << self
         def call(object, env)
           MultiXml.parse(object)
+        rescue MultiXml::ParseError
+          # handle XML parsing errors via the rescue handlers or provide error message
+          raise Grape::Exceptions::InvalidMessageBody, 'application/xml'
         end
       end
     end